Add showbook book bookform js files
